feat: add debugging documentation and enhance applicant form logging
- Introduce a new markdown file for debugging the unmask button issue on the /board page.
- Enhance logging in the `handleCreateCandidate` function to trace form state and validation.
- Update imports to reflect changes in the applicant interface structure.
- Refactor error handling and button state management for improved user feedback.
- Remove deprecated CV interface and consolidate applicant-related types for consistency.
