+++ frontend/debug-unmask-button.md
@@ -1,0 +1,38 @@
+# Debugging the Unmask Button Issue
+
+## Current Investigation
+
+The unmask button is not triggering the workflow when pressed on the /board page after filling:
+- LinkedIn link
+- GitHub link  
+- CV profile file upload
+
+## Findings So Far
+
+1. ✅ **API Endpoint Works**: Tested `/api/applicants` directly with curl - it responds correctly
+2. ✅ **Form Validation Logic**: `isFormValid` checks `(linkedinUrl.trim() || cvFile) && !isCreating && !isLoading`
+3. ✅ **Button Handler**: `handleCreateCandidate` function exists and calls `createApplicant`
+4. ✅ **Context Method**: `createApplicant` in ApplicantContext posts to `/api/applicants`
+
+## Potential Issues to Check
+
+1. **Form Validation**: Is `isFormValid` returning `false`?
+2. **Button State**: Is the button disabled due to validation?
+3. **JavaScript Errors**: Are there console errors preventing execution?
+4. **Network Issues**: Is the fetch request failing silently?
+5. **Environment Variables**: Are required env vars (GROQ_API_KEY, BRIGHTDATA_API_KEY) set?
+
+## Debug Steps Needed
+
+1. Add console.log to `handleCreateCandidate` to see if it's called
+2. Check browser console for JavaScript errors
+3. Add logging to `createApplicant` to see where it fails
+4. Verify form validation state
+5. Check network tab for failed requests
+
+## Quick Fix Suggestions
+
+1. Add error boundary around the form
+2. Add better error logging
+3. Add visual feedback for button states
+4. Check if ApplicantProvider is wrapping the component correctly
