+++ frontend/src/lib/interfaces/index.ts
@@ -1,5 +1,8 @@
-// CV-related interfaces and enums
-export * from './cv';
-
-// Applicant-related interfaces
+// Applicant-related interfaces (includes CV/profile types)
 export * from './applicant';
+
+// GitHub-related interfaces
+export * from './github';
+
+// Analysis-related interfaces
+export * from './analysis';
